+++ lib/releaf/rspec/matchers/cache_instance_method.rb
@@ -1,0 +1,6 @@
+RSpec::Matchers.define :cache_instance_method do |method_name|
+
+  match do |actual|
+    described_class.cached_instance_methods.include?(method_name)
+  end
+end