+++ releaf-core/spec/builders/releaf/builders/table_builder_spec.rb
@@ -407,7 +407,7 @@ describe Releaf::Builders::TableBuilder, type: :class do
       allow(subject).to receive(:column_value).with(resource, :birth_date)
         .and_return(value)
 
-      expect(I18n).to receive(:l).with(value, format: :default, default: "%Y-%m-%d")
+      expect(I18n).to receive(:l).with(value, format: :default)
         .and_call_original
 
       expect(subject.format_date_content(resource, :birth_date)).to eq("2012-12-29")
@@ -420,7 +420,7 @@ describe Releaf::Builders::TableBuilder, type: :class do
       allow(subject).to receive(:column_value).with(resource, :created_at)
         .and_return(value)
 
-      allow(I18n).to receive(:l).with(value, format: :default, default: "%Y-%m-%d %H:%M:%S")
+      allow(I18n).to receive(:l).with(value, format: "%Y-%m-%d %H:%M")
         .and_return("2012-12-29 17:12:07")
 
       expect(subject.format_datetime_content(resource, :created_at)).to eq("2012-12-29 17:12:07")