+++ releaf-i18n_database/lib/releaf/i18n_database/backend.rb
@@ -4,12 +4,15 @@ module Releaf
   module I18nDatabase
     class Backend
 
-      include ::I18n::Backend::Base, ::I18n::Backend::Flatten
+      include ::I18n::Backend::Base
+      include ::I18n::Backend::Flatten
+      include ::I18n::Backend::Pluralization
+
       UPDATED_AT_KEY = 'releaf.i18n_database.translations.updated_at'
       DEFAULT_CONFIG = {
         translation_auto_creation: true,
         translation_auto_creation_patterns: [/.*/],
-        translation_auto_creation_exclusion_patterns: [/^attributes\./]
+        translation_auto_creation_exclusion_patterns: [/^attributes\./, /^i18n\./]
       }
       attr_accessor :translations_cache
 
@@ -18,9 +21,11 @@ module Releaf
       end
 
       def self.locales_pluralizations
-        Releaf.application.config.all_locales.map do|locale|
-          TwitterCldr::Formatters::Plurals::Rules.all_for(locale) if TwitterCldr.supported_locale?(locale)
-        end.flatten.uniq.compact
+        keys = Releaf.application.config.all_locales.map{ |locale| I18n.t(:'i18n.plural.keys', locale: locale) }.flatten
+        # always add zero as it skipped for some locales even when there is zero form (lv for example)
+        keys << :zero
+
+        keys.uniq
       end
 
       def self.configure_component